Minitalk qui marche moin bien qu'avant

+++ minitalk/client.c
@@ -6,12 +6,11 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 13:12:56 by codespace         #+#    #+#             */
-/*   Updated: 2023/12/15 15:03:56 by codespace        ###   ########.fr       */
+/*   Updated: 2023/12/18 18:23:44 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
-#include <stdio.h>
 
 void	send_signal(int pid, unsigned char character)
 {
@@ -24,29 +23,30 @@ void	send_signal(int pid, unsigned char character)
 	{
 		i--;
 		tmp = character >> i;
-		if (tmp % 2)
+		if (tmp % 2 )
 			kill(pid, SIGUSR1);
-		else
+		else 
 			kill(pid, SIGUSR2);
-		usleep(10000);
+		usleep(1000);
 	}
 }
 
 int	main(int argc, char **argv)
 {
-	(void)argc;
 	int i;
 	int j;
 	i = 0;
 	j = 0;
 
-	while (argv[2][i])
-		i++;
-	while (i > j)
+	if (argc == 3)
 	{
-		send_signal(atoi(argv[1]), argv[2][j]);
-		j++;
+		while (argv[2][i])
+			i++;
+		while (i > j)
+		{
+			send_signal(atoi(argv[1]), argv[2][j]);
+			j++;
+		}
+		return (1);
 	}
-	// write(1, "\n", 1);
-	return (0);
 }
